Populate season_statistics class
