+++ runner.rb
@@ -11,5 +11,6 @@ locations = {
 }
 
 stat_tracker = StatTracker.from_csv(locations)
+season_tracker = StatTracker.from_csv(locations)
 
 require 'pry'; binding.pry
